refactory: proof subtitle
