+++ public/locales/en/home.json
@@ -1,0 +1,27 @@
+{
+  "subtitleProofread": "Subtitle Proofreading",
+  "videoNotFound": "Video file not found",
+  "playbackControls": "Playback Controls",
+  "rewind5Seconds": "Rewind 5 seconds",
+  "previousSubtitle": "Previous subtitle",
+  "play": "Play",
+  "pause": "Pause",
+  "nextSubtitle": "Next subtitle",
+  "forward5Seconds": "Forward 5 seconds",
+  "playbackSpeed": "Playback speed",
+  "currentSubtitle": "Current Subtitle",
+  "noCurrentSubtitle": "No current subtitle",
+  "fileInfo": "File Information",
+  "unknown": "Unknown",
+  "duration": "Duration",
+  "subtitleStats": "Subtitle Statistics",
+  "total": "Total",
+  "translated": "Translated",
+  "completionRate": "Completion rate",
+  "originalSubtitle": "Original subtitle",
+  "translatedSubtitle": "Translated subtitle",
+  "cancel": "Cancel",
+  "save": "Save",
+  "subtitleSavedSuccess": "Subtitles saved successfully",
+  "saveFailed": "Failed to save subtitles"
+}
