+++ public/locales/zh/home.json
@@ -1,0 +1,27 @@
+{
+  "subtitleProofread": "字幕校对",
+  "videoNotFound": "未找到视频文件",
+  "playbackControls": "播放控制",
+  "rewind5Seconds": "后退5秒",
+  "previousSubtitle": "上一句",
+  "play": "播放",
+  "pause": "暂停",
+  "nextSubtitle": "下一句",
+  "forward5Seconds": "前进5秒",
+  "playbackSpeed": "播放速度",
+  "currentSubtitle": "当前字幕",
+  "noCurrentSubtitle": "无当前字幕",
+  "fileInfo": "文件信息",
+  "unknown": "未知",
+  "duration": "时长",
+  "subtitleStats": "字幕统计",
+  "total": "总数",
+  "translated": "已翻译",
+  "completionRate": "完成率",
+  "originalSubtitle": "原文字幕",
+  "translatedSubtitle": "翻译字幕",
+  "cancel": "取消",
+  "save": "保存",
+  "subtitleSavedSuccess": "字幕保存成功",
+  "saveFailed": "保存失败"
+}
