+++ renderer/public/locales/en/home.json
@@ -86,5 +86,25 @@
   "originalText": "Original Text",
   "translatedText": "Translated Text",
   "subtitleSavedSuccess": "Subtitle saved successfully",
-  "saveFailed": "Save failed"
+  "saveFailed": "Save failed",
+  "translatedSubtitle": "Translated Subtitle",
+  "videoNotFound": "Video file not found",
+  "playbackControls": "Playback Controls",
+  "rewind5Seconds": "Rewind 5 seconds",
+  "previousSubtitle": "Previous subtitle",
+  "play": "Play",
+  "pause": "Pause",
+  "nextSubtitle": "Next subtitle",
+  "forward5Seconds": "Forward 5 seconds",
+  "playbackSpeed": "Playback speed",
+  "currentSubtitle": "Current Subtitle",
+  "noCurrentSubtitle": "No current subtitle",
+  "fileInfo": "File Information",
+  "unknown": "Unknown",
+  "duration": "Duration",
+  "subtitleStats": "Subtitle Statistics",
+  "total": "Total",
+  "translated": "Translated",
+  "completionRate": "Completion rate",
+  "originalSubtitle": "Original subtitle"
 }
